20_advanced/01_rest_api/final/src/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-filter==24.1
python-decouple==3.8
Pillow==10.3.0
whitenoise==6.6.0
